+++ GeneralsMD/Code/GameEngine/Source/GameLogic/Object/Update/WaveGuideUpdate.cpp
@@ -802,7 +802,7 @@ UpdateSleepTime WaveGuideUpdate::update( void )
 	}
 
 	// every half second we try to play a random splash sound
-	if( TheGameLogic->getFrame() - m_splashSoundFrame > LOGICFRAMES_PER_SECOND / 2.0f )
+	if( TheGameLogic->getFrame() - m_splashSoundFrame > static_cast<float>(LOGICFRAMES_PER_SECOND) / 2.0f )
 	{
 
 		// mark that we've had the opportunity to play
